cmd/gvisor-website: add /rebuild and cron
In order to ensure that the system call descriptions and other
documentation that will be sourced from the main repository remains up
to date, add a hook that allows the site to submit a rebuild to Google
Cloud Builder. This can be triggered manually by an admin, or will be
triggered every 24 hours by a new cron configuration.